requirements.txt
flake8==4.0.1
mccabe==0.6.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
